+++ deploy/qemu/vm-bootstrap/integrations.sh
@@ -1,2 +1,0 @@
-#!/bin/sh
-exec sh /mnt/korus/deploy/qemu/vm-bootstrap/run-ansible-local.sh integrations