+++ deploy/qemu/vm-bootstrap/web.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-# QEMU web VM bootstrap (Ansible). See deploy/qemu/README.md
-exec sh /mnt/korus/deploy/qemu/vm-bootstrap/run-ansible-local.sh web